+++ app/resources/numismatics/issues/numismatics/issue_change_set.rb
@@ -57,6 +57,7 @@ module Numismatics
     property :downloadable, multiple: false, require: true, default: "public"
     property :rights_statement, multiple: false, required: true, default: RightsStatements.no_known_copyright, type: ::Types::URI
     property :rights_note, multiple: false, required: false
+    property :state, multiple: false, required: true, default: "complete"
 
     # Virtual Attributes
     property :files, virtual: true, multiple: true, required: false